requirements.txt
# Core
django==4.2.8
python-dotenv==1.0.0
django-snowflake==5.1
djangorestframework==3.15.2

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0
django-debug-toolbar==4.2.0
